use root sidedrawer, clean up
