--- app/views/main-page/main-content.xml
@@ -1,47 +1,0 @@
-<!-- <GridLayout id="content-root">
-        <GridLayout id="intro-elements" onTap={enter}>
-            <GridLayout id="intro-background" class="intro-background-intro" originY="0" />
-
-            <GridLayout id="intro-logo-bg" class="intro-logo-bg-intro" iosOverflowSafeArea="false" backgroundColor="#3C5AFD" width="93" height="93" horizontalAlignment="center" verticalAlignment="center" borderRadius="20" />
-            <GridLayout id="intro-logo-n" class="intro-logo-n-intro" iosOverflowSafeArea="false" backgroundImage="res://logo_blue_bg" width="93" height="93" horizontalAlignment="center" verticalAlignment="center" />
-            <GridLayout id="intro-logo-ns" class="intro-logo-ns-intro" iosOverflowSafeArea="false" backgroundImage="res://logo_text" width="199" height="31" horizontalAlignment="center" verticalAlignment="center" margin="160 0 0 0" />
-
-            <Label id="intro-text-one"
-                class="intro-text-one-intro"
-                text="Build truly&#xA;native apps with&#xA;JavaScript"
-                fontSize="37"
-                horizontalAlignment="center"
-                verticalAlignment="center"
-                textWrap="true"
-                color="white"
-                textAlignment="center" />
-            <Label id="intro-text-two"
-                class="intro-text-two-intro"
-                text="Develop native cross platform&#xA;apps from a single code base"
-                fontSize="19"
-                horizontalAlignment="center"
-                verticalAlignment="center"
-                textWrap="true"
-                color="#8DA1AB"
-                textAlignment="center" />
-
-            <Button id="intro-get-started"
-                class="intro-get-started-intro"
-                text="GET STARTED"
-                backgroundColor="#1DBE67"
-                color="white"
-                borderRadius="25"
-                height="50"
-                width="234"
-                fontSize="18"
-                tap="tapGetStarted" />
-            <Label id="intro-version"
-                class="intro-version-intro"
-                text="version 4.0.0"
-                fontSize="14"
-                horizontalAlignment="center"
-                verticalAlignment="center"
-                color="white"
-                textAlignment="center" />
-        </GridLayout>
-    </GridLayout> -->
